adjusting data model, implemennt ux for multiple product bundling

+++ features/merchant/product_bundle/src/main/java/com/tokopedia/product_bundle/multiple/presentation/model/ProductBundleDetail.kt
@@ -3,7 +3,7 @@ package com.tokopedia.product_bundle.multiple.presentation.model
 class ProductBundleDetail(
         val productImageUrl: String = "",
         val productName: String = "",
-        val originalPrice: Int = 0,
-        val bundlePrice: Int = 0,
-        val discountAmount: Int = 0
+        val originalPrice: Double = 0.0,
+        val bundlePrice: Double = 0.0,
+        val discountAmount: Double = 0.0
 )
